pyproject.toml
[project]
name = "vqa-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tyro",
    "opencv-python",
    "static_error_handler",
    "pillow",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vqa_pipeline"]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"
